Fix internal peer pool enqueue errors
ThaliManager.stop() should call stop method of every dependency
(ThaliMobile, pullReplication etc.). But it was done in a wrong order so
it was possible to get enqueue error in the following scenario:

1. We have a peer connected to other peers and probably in the middle of
   live replication
2. We decide to shut down everything (call ThaliManager.stop())
3. The first thing stop method does is it stops a
   thaliPeerPoolInterface. And when thaliPeerPoolInterface stops it kills
   every enqueued action. Since we are in the middle of active usage
   we probably have some replication and notification actions in the
   queue
4. When NotificationActions is unexpectedly killed, NotificationClient
   tries to restart it after some timeout. And "restart action" means
   that notification client will create new notification action and
   enqueue it into the peer pool manager. As soon as timeout passes,
   client enqueues this new action into stopped peer pool. And we have
   exception.

Another possible scenario is when we are getting notifications from
other peers while out peer pool is stopping, because ThaliMobile and
ThaliNotificationClient are still active.

It is fixed by stopping everything simultaneously. We don't have to wait
for peer pool before stopping everything else. This way notification
client won't restart failed actions and won't react to new
notifications.