Fix wifi network changes handling
Previously bssidName:<MAC address> -> bssidName:null was considered as
simple BSSID change and wifiInfrastructure was restarting ssdp client
and server. SSDP start always fails when bssidName is null because it
means that we are not connected to any access point and couldn't
advertise anything.

In reality bssid change from/to null should be handled the same way as
enabling/disabling WiFi on device.